+++ src/main/java/cc/cassian/rrv/common/builtin/villager/VillagerServerRecipe.java
@@ -105,7 +105,7 @@ public class VillagerServerRecipe implements ReliableServerRecipe {
 
 	public SlotContent offerStacks(VillagerTrade trade) {
 		VillagerTradeAccessor tradeAccessor = (VillagerTradeAccessor) trade;
-		AtomicReference<ItemStack> stack = new AtomicReference<>(RrvUtil.decodeTemplate(tradeAccessor.getGives()));
+		AtomicReference<ItemStack> stack = new AtomicReference<>((tradeAccessor.getGives().create()));
 		ServerRecipeManager.INSTANCE.setWorkaroundFlag(true);
 		tradeAccessor.getGivenItemModifiers().forEach(modifier -> {
 			if (modifier instanceof ExplorationMapFunction || modifier instanceof FilteredFunction) return; // utter bodge - fixes the map item getting entirely voided